requirements.txt
Django==1.7.10
Markdown==2.6.5
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.11.0
django-toolbelt==0.0.1
djangorestframework==3.3.1
gunicorn==19.4.1
psycopg2==2.6.1
static3==0.6.1
wsgiref==0.1.2
